Extract utils from middleware
